Removing useCallback call to retrieve deliveries data
The useCallback was being called a lot of times due to unecessary situations. You need to care with the dependencies array. 

Instead of using the useCallback hook, you can create the async function inside of the useEffect hook, and then put the right dependencies to run it. 
Since it's going to be a list and the data will have to appear on the first mount of the component, i putted a empty dependencies array, but you can change according to what you want to achieve.